refactor(cli): split execution modules
